Move walk_dir() from fawltydeps.utils to tests.utils
traverse_projects.find_sources() recently replaced its use of walk_dir()
with the more advanced DirectoryTraversal.With that there are no
remaining users of walk_dir() in the fawltydeps package. There are still
a few uses in our test code, so we move walk_dir() to tests.utils.
